Fix wrong MapViewModel used in updateAllDistancesAndFindClosest
